added access controll

+++ backend/src/routers/item_preset_router.rs
@@ -1,8 +1,16 @@
-
 use rocket::{form::FromForm, http::Status, response::status::Custom, serde::json::Json, State};
+use serde::{Deserialize, Serialize};
+use crate::controller::{cstat, new_cstst};
 use crate::{controller::item_preset_controller::ItemPresetController, model::ItemPreset};
-
+use crate::controller::inventory_controller::InventoryController;
 use super::ttjhe;
+
+
+#[derive(Serialize, Deserialize)]
+pub struct GetItemPresetReturn{
+    item_presets: Vec<ItemPreset>
+}
+
 #[derive(FromForm)]
 pub struct ItemPresetUUIDParams {
     item_preset_uuid: String
@@ -16,16 +24,35 @@ pub struct ItemModifyParams {
     description: Option<String>,
     item_type: Option<String>
 }
-//TODO: Access controll
+
+fn has_access_to(searched_item_preset: String, inventories: Vec<String>, inv_con: &State<InventoryController>) -> Result<bool, cstat> {
+    let mut has_access = false;
+    for i in inventories {
+        if inv_con.item_exits(i, searched_item_preset.clone())? {
+            has_access = true
+        }
+    }
+    return Ok(has_access)
+}
+
 #[get("/itemPreset?<params..>")]
-pub async fn get_item_preset(params: ItemPresetUUIDParams,  user: super::AuthenticatedUser,
-        ipc_con: &State<ItemPresetController>) -> Result<Json<ItemPreset>, Custom<&'static str>> {
-        ttjhe(ipc_con.get_item_preset(params.item_preset_uuid), Status::InternalServerError)
+pub async fn get_item_preset(params: ItemPresetUUIDParams,  user: super::AuthenticatedUser, ipc_con: &State<ItemPresetController>,
+        inv_con: &State<InventoryController>) -> Result<Json<ItemPreset>, cstat> {
+    let invs = inv_con.get_all_inventories_ids(user.user_id)?;
+    
+    if !has_access_to(params.item_preset_uuid.clone(), invs, inv_con)? {
+        return Err(new_cstst(Status::Forbidden, "No access"));
+    }
+    ttjhe(ipc_con.get_item_preset(params.item_preset_uuid), Status::InternalServerError)
 }
 
 #[patch("/itemPreset/modify?<params..>")]
 pub async fn modify_item_preset(params: ItemModifyParams,  user: super::AuthenticatedUser,
-        ipc_con: &State<ItemPresetController>) -> Result<Status, Custom<&'static str>> {
+        ipc_con: &State<ItemPresetController>, inv_con: &State<InventoryController>) -> Result<Status, Custom<&'static str>> {
+    let invs = inv_con.get_all_inventories_ids_with_read_access(user.user_id)?;
+    if !has_access_to(params.item_preset_uuid.clone(), invs, inv_con)? {
+        return Err(new_cstst(Status::Forbidden, "No access"));
+    }
     match ipc_con.edit_item_preset(params.item_preset_uuid, params.name, params.price, params.description, params.item_type){
         Ok(_res) => Ok(Status::NoContent),
         Err(e) => Err(Custom(
@@ -37,7 +64,12 @@ pub async fn modify_item_preset(params: ItemModifyParams,  user: super::Authenti
 
 #[patch("/itemPreset/delete?<params..>")]
 pub async fn delete_item_preset(params: ItemPresetUUIDParams,  user: super::AuthenticatedUser,
-        ipc_con: &State<ItemPresetController>) -> Result<Status, Custom<&'static str>> {
+        ipc_con: &State<ItemPresetController>, inv_con: &State<InventoryController>) -> Result<Status, Custom<&'static str>> {
+    let invs = inv_con.get_all_inventories_ids_with_read_access(user.user_id)?;
+    if !has_access_to(params.item_preset_uuid.clone(), invs, inv_con)? {
+        return Err(new_cstst(Status::Forbidden, "No access"));
+    }
+
     match ipc_con.delete_item_preset(params.item_preset_uuid) {
         Ok(_res) => Ok(Status::NoContent),
         Err(e) => Err(Custom(
@@ -48,7 +80,16 @@ pub async fn delete_item_preset(params: ItemPresetUUIDParams,  user: super::Auth
 }
 
 #[patch("/itemPreset/all")]
-pub async fn get_all_item_presets(user: super::AuthenticatedUser) -> &'static str {
-    // return all inventories
-    "Hello, Rocket with async!"
+pub async fn get_all_item_presets(user: super::AuthenticatedUser, inv_con: &State<InventoryController>,
+        ipc_con: &State<ItemPresetController>) -> Result<Json<GetItemPresetReturn>, cstat> {
+    let mut ips: Vec<ItemPreset> = Vec::new();
+    let invs = inv_con.get_all_inventories_ids(user.user_id)?;
+    for i in invs {
+        ips.extend(ipc_con.get_item_preset_in_inventory(i)?)
+    }
+    Ok(Json(
+        GetItemPresetReturn {
+            item_presets: ips
+        }
+    ))
 }
